mfi_mpower: Disable redirects for login POST request
The login POST requests results in a chain of 302 redirects where
the last one loads the main web UI. This last request sometimes
hangs not allowing python3 requests to cleanly exit. This gives the
impression that the login did not work at all.

By disabling the redirects in the login POST message we bypass the
redirect chain that results in the web UI load attempt while still
getting an authenticated cookie in the response.

Signed-off-by: Maciej Grela <m.grela@samsung.com>